% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizerocs.R
\name{visualize_full_roc}
\alias{visualize_full_roc}
\title{Simulate and Plot ROC Curves for Two Groups}
\usage{
visualize_full_roc(
  mean_signal_g1 = 1,
  mean_signal_g2 = 0.58,
  n_studied = 5,
  n_new = 5,
  n_g1 = 10000,
  sd_signal_g1 = 1,
  sd_signal_g2 = 1,
  sd_noise_g1 = 1,
  sd_noise_g2 = 1,
  mean_noise_g1 = 0,
  mean_noise_g2 = 0,
  rho = 0,
  seed = 2794
)
}
\arguments{
\item{mean_signal_g1}{Mean of the signal distribution for Group 1.}

\item{mean_signal_g2}{Mean of the signal distribution for Group 2.}

\item{n_studied}{Number of studied items per participant.}

\item{n_new}{Number of new items per participant.}

\item{n_g1}{Number of participants (default = 10000 for stable AUC).}

\item{sd_signal_g1}{Standard deviation of the signal for Group 1.}

\item{sd_signal_g2}{Standard deviation of the signal for Group 2.}

\item{sd_noise_g1}{Standard deviation of the noise for Group 1.}

\item{sd_noise_g2}{Standard deviation of the noise for Group 2.}

\item{mean_noise_g1}{Mean of the noise for Group 1.}

\item{mean_noise_g2}{Mean of the noise for Group 2.}

\item{rho}{Correlation between the two groups (default = 0).}

\item{seed}{Random seed for reproducibility.}
}
\value{
A list containing:
\itemize{
\item \code{roc_g1}: ROC object for Group 1
\item \code{roc_g2}: ROC object for Group 2
\item \code{plot_g1}: ROC plot for Group 1
\item \code{plot_g2}: ROC plot for Group 2
}
}
\description{
This function simulates ROC data for two groups, generates confidence ratings,
computes ROC curves, and returns their plots with AUC.
}
\examples{
# Get ROC curves and AUCs for default parameters
result <- visualize_full_roc()
# To get ROC curves and AUC for single group use group 1 statistics

Visualize ROC curves and AUC for parameters of interest
visualize_full_roc(mean_signal_g1 = 1, mean_signal_g2 = 0.58,
n_studied = 5, n_new = 5,
n_g1 = 10000, # Is high to get precise estimate of ROC curve and AUC
sd_signal_g1 = 1, sd_signal_g2 = 1,
sd_noise_g1 = 1, sd_noise_g2 = 1,
mean_noise_g1 = 0, mean_noise_g2 = 0,
rho = 0, seed = 2794)
}
